// src/app/api/model/budget.ts
/**
 * Personal Finance Management API
 *
 * Contact: dragisa@yahoo.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Category } from './category';


export interface Budget { 
    id?: number;
    username?: string;
    amount?: number;
    month?: number;
    year?: number;
    category?: Category;
    createdAt?: string;
    updatedAt?: string;
}